requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

# Core scientific stack
numpy
pillow
tqdm
matplotlib
ipython
datasets>=2.19

# Hugging Face ecosystem
transformers>=4.51,<5
tokenizers>=0.15
accelerate>=0.33
huggingface-hub>=0.34
sentencepiece
tiktoken
protobuf

# Qwen utilities
qwen-vl-utils>=0.0.11

# PyTorch + vision/audio (CUDA 12.1 builds)
torch==2.4.1+cu121
torchvision==0.19.1+cu121
torchaudio==2.4.1+cu121

# Performance helpers
bottleneck>=1.3.6

scipy==1.14.*
scikit-learn==1.5.*
